requirements.txt
# Backend dependencies

# Core web framework
fastapi
uvicorn[standard] # ASGI server

# Database (SQLite)
sqlalchemy
aiosqlite # Async driver for SQLite
# Removed: asyncpg, psycopg2-binary
alembic # For database migrations (still useful even for SQLite)

# Pydantic (already included with FastAPI, but explicit is fine)
pydantic
pydantic-settings # For loading settings from .env

# AI - Langchain & Groq
langchain
langgraph
langchain_groq
groq # Groq Python client
langchain_core
langchain_community

# Other utilities
python-dotenv # To load .env files
python-multipart # For form data if needed

# CORS
fastapi-cors # If using the specific middleware, otherwise FastAPI has built-in support
